move logic from `LandingPage` to `ProductSearchForm`
