+++ web/src/components/molecules/ProductSearchForm.test.tsx
@@ -1,50 +1,38 @@
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
 import { render, screen } from "@testing-library/react";
 import { ProductSearchForm } from "./ProductSearchForm";
 import userEvent from "@testing-library/user-event";
+import { http, HttpResponse, server } from "../../../tests/setup-server";
 
 describe("ProductSearchForm", () => {
-  it("displays label", async () => {
-    // Act
-    render(<ProductSearchForm onSubmit={() => { }} />);
-
-    // Assert
-    const label = screen.getByText("Product URL");
-    expect(label).toBeVisible();
-  });
-
-  it("displays text box", async () => {
-    // Act
-    render(<ProductSearchForm onSubmit={() => { }} />);
-
-    // Assert
-    const input = screen.getByRole("textbox");
-    expect(input).toBeVisible();
-  });
-
-  it("displays button", async () => {
-    // Act
-    render(<ProductSearchForm onSubmit={() => { }} />);
-
-    // Assert
-    const button = screen.getByRole("button", { name: "Search" });
-    expect(button).toBeVisible();
-  });
-
-  it("triggers onSubmit when user clicks the button", async () => {
+  it("displays Product Detail after search", async () => {
     // Arrange
-    const user = userEvent.setup();
+    server.use(
+      http.get("/api/products/getByUrl", ({ request }) => {
+        const url = new URL(request.url);
+        const productUrl = url.searchParams.get("url");
+        if (productUrl != "https://google.com") {
+          return new HttpResponse(null, { status: 404 });
+        }
 
-    const mockFunction = vi.fn();
-    render(<ProductSearchForm onSubmit={mockFunction} />);
+        return HttpResponse.json({
+          sku: "123",
+          name: "A dummy product",
+        });
+      }),
+    );
+
+    const user = userEvent.setup();
+    render(<ProductSearchForm />);
 
     // Act
-    const textbox = screen.getByRole("textbox");
+    const textbox = screen.getByRole("textbox", { name: "Product URL" });
     await user.type(textbox, "https://google.com");
-    const button = screen.getByRole("button");
+    const button = screen.getByRole("button", { name: "Search" });
     await user.click(button);
 
     // Assert
-    expect(mockFunction).toHaveBeenCalledWith("https://google.com");
+    const product = screen.getByText("A dummy product");
+    expect(product).toBeVisible();
   });
 });
